+++ api/src/controllers/category/index.ts
@@ -1,0 +1,3 @@
+export * from "./create.category.controller";
+export * from "./get.category.controller";
+export * from "./delete.category.controller";
